pyproject.toml
[project]
name = "hbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.12.0",
    "google>=3.0.0",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "jsondb-hakimifr>=0.1.0",
    "kurigram>=2.2.15",
    "tgcrypto>=1.2.5",
    "uvloop>=0.22.1",
]
